requirements.txt
asgiref==3.11.1
click==8.3.1
colorama==0.4.6
cssbeautifier==1.15.4
Django==6.0.3
django-cors-headers==4.9.0
django-extensions==4.1
django-pwa==2.0.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
djlint==1.36.4
EditorConfig==0.17.1
et_xmlfile==2.0.0
ExifRead==3.5.1
Faker==40.8.0
geographiclib==2.1
geopy==2.4.1
jsbeautifier==1.15.4
json5==0.13.0
numpy==2.4.3
openpyxl==3.1.5
pandas==3.0.1
pathspec==1.0.4
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.2.28
six==1.17.0
sqlparse==0.5.5
tqdm==4.67.3
